add feedback gate and crm stuff

+++ src/main/java/com/reputul/backend/controllers/BusinessController.java
@@ -98,13 +98,13 @@ public class BusinessController {
 
         double avgRating = reviewRepo.findAverageRatingByBusinessId(id) != null
                 ? reviewRepo.findAverageRatingByBusinessId(id) : 0.0;
-        int total = reviewRepo.countByBusinessId(id);
+        long total = reviewRepo.countByBusinessId(id);
         List<Review> recent = reviewRepo.findTop1ByBusinessIdOrderByCreatedAtDesc(id);
         String comment = recent.isEmpty() ? "No reviews yet." : recent.get(0).getComment();
 
         String badge = business.getBadge() != null ? business.getBadge() : "Unranked";
 
-        ReviewSummaryDto summary = new ReviewSummaryDto(avgRating, total, comment, badge);
+        ReviewSummaryDto summary = new ReviewSummaryDto(avgRating, (int) total, comment, badge);
         return ResponseEntity.ok(summary);
     }
